Some cleanups.
I realized that there is a poison-pill strategy of giving scores to lots
of unpopular candidates.  This would make your ballot not be considered
as truncating.  The consequence is that votes are rebalanced less efficiently --
we want to ensure that voters who aren't spreading their approval over as
many viable candidates get their ballots used up first, as an anti-truncation
incentive.

We need to ensure that the truncation penalty of using up truncated ballots
first is used in realistic cases.

Therefore, we remove candidates with less than half a quota of non-zero
total approval before starting.